Move section into cv folder and wrap some tests around it
